microservices/ocrService/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.2.1
pytesseract==0.3.13
textdistance==4.6.3
Werkzeug==3.1.3
